requirements.txt
pandas>=2.1
numpy>=1.26
pyarrow>=15.0
pydantic>=2.5
pyyaml>=6.0
scikit-learn>=1.4
fastapi>=0.110
uvicorn>=0.27
tqdm>=4.66
faiss-cpu>=1.8
networkx>=3.2
python-dotenv>=1.0
sentence-transformers>=3.0
requests
pytest
streamlit==1.40.0